25-02-08 routes 분리 및 tarot_service & data 파일 분리 수정
