+++ app/__init__.py
@@ -1,4 +1,4 @@
-from flask import Flask, session, redirect
+from flask import Flask
 from app.models import db, User
 from flask_migrate import Migrate
 from flask_wtf.csrf import CSRFProtect
@@ -15,16 +15,21 @@ def create_app():
     app.config.from_object('app.config.Config')
     
     csrf.init_app(app)
+
     db.init_app(app)
     migrate.init_app(app, db)
-    login_manager.init_app(app)
     socketio.init_app(app)
 
+    login_manager.init_app(app)
+    login_manager.login_view = "auth.login"  
+    
     @login_manager.user_loader
     def load_user(user_id):
         return User.query.get(user_id)
     
-    from app.routes import bp
-    app.register_blueprint(bp)
+    from app.routes import auth_bp, chat_bp, ws_bp
+    app.register_blueprint(auth_bp, url_prefix='/')
+    app.register_blueprint(chat_bp, url_prefix='/chat')
+    app.register_blueprint(ws_bp, url_prefix='/ws')
 
     return app
