+++ app/data/tarot_meaning.py
@@ -777,117 +777,3 @@ TAROT_CARD_MEANINGS = {
         "당신은 보호받고 있으며, 깨달음을 얻었다."
     ]
 }
-
-
-TAROT_CARD_INFO = {
-    "The Fool": {
-        "number": 0,
-        "keywords": ["new beginnings", "adventure", "spontaneity"],
-        "description": "The Fool symbolizes the start of a new journey, full of potential and opportunities. It encourages you to take risks and embrace the unknown."
-    },
-    "The Magician": {
-        "number": 1,
-        "keywords": ["power", "skill", "manifestation"],
-        "description": "The Magician represents personal power, resourcefulness, and the ability to manifest your goals. It is a card of action and creation."
-    },
-    "The High Priestess": {
-        "number": 2,
-        "keywords": ["intuition", "mystery", "inner wisdom"],
-        "description": "The High Priestess signifies intuition, mystery, and deep understanding. It encourages you to trust your instincts and seek hidden truths."
-    },
-    "The Empress": {
-        "number": 3,
-        "keywords": ["abundance", "nurturing", "creativity"],
-        "description": "The Empress embodies fertility, abundance, and creativity. It suggests nurturing relationships and fostering growth."
-    },
-    "The Emperor": {
-        "number": 4, 
-        "keywords": ["authority", "structure", "stability"],
-        "description": "The Emperor symbolizes authority, structure, and discipline. It calls for leadership and strategic planning."
-    },
-    "The Hierophant": {
-        "number": 5, 
-        "keywords": ["tradition", "spiritual guidance", "conformity"],
-        "description": "The Hierophant represents tradition, spiritual guidance, and learning. It encourages seeking wisdom from established systems or mentors."
-    },
-    "The Lovers": {
-        "number": 6, 
-        "keywords": ["love", "choices", "harmony"],
-        "description": "The Lovers signify relationships, harmony, and important choices. It highlights the balance between emotions and logic."
-    },
-    "The Chariot": {
-        "number": 7,
-        "keywords": ["determination", "willpower", "victory"],
-        "description": "The Chariot symbolizes determination, focus, and overcoming challenges. It urges you to stay on course and keep control."
-    },
-    "Strength": {
-        "number": 8,
-        "keywords": ["courage", "compassion", "inner strength"],
-        "description": "Strength represents courage, patience, and compassion. It reminds you to approach challenges with calm and resilience."
-    },
-    "The Hermit": {
-        "number": 9,
-        "keywords": ["introspection", "solitude", "wisdom"],
-        "description": "The Hermit signifies a time for introspection and seeking inner wisdom. It encourages solitude and self-discovery."
-    },
-    "The Wheel of Fortune": {
-        "number": 10,
-        "keywords": ["change", "destiny", "cycles"],
-        "description": "The Wheel of Fortune represents the cycles of life, changes, and unexpected events. It suggests adapting to new circumstances."
-    },
-    "Justice": {
-        "number": 11,
-        "keywords": ["truth", "fairness", "accountability"],
-        "description": "Justice symbolizes fairness, truth, and accountability. It calls for honest evaluation and responsible actions."
-    },
-    "The Hanged Man": {
-        "number": 12,
-        "keywords": ["sacrifice", "perspective", "letting go"],
-        "description": "The Hanged Man signifies a pause, a shift in perspective, and the need to let go of control to gain insight."
-    },
-    "Death": {
-        "number": 13,
-        "keywords": ["transformation", "endings", "rebirth"],
-        "description": "Death represents transformation, endings, and new beginnings. It signals the closure of a chapter and the start of another."
-    },
-    "Temperance": {
-        "number": 14,
-        "keywords": ["balance", "moderation", "healing"],
-        "description": "Temperance symbolizes harmony, moderation, and healing. It encourages patience and finding balance in your life."
-    },
-    "The Devil": {
-        "number": 15,
-        "keywords": ["temptation", "bondage", "materialism"],
-        "description": "The Devil represents temptation, addiction, and materialism. It calls for breaking free from unhealthy attachments."
-    },
-    "The Tower": {
-        "number": 16,
-        "keywords": ["upheaval", "revelation", "destruction"],
-        "description": "The Tower signifies sudden upheaval, revelations, and the breaking down of old structures. It paves the way for new growth."
-    },
-    "The Star": {
-        "number": 17,
-        "keywords": ["hope", "inspiration", "renewal"],
-        "description": "The Star represents hope, inspiration, and renewal. It is a card of positivity and dreams coming true."
-    },
-    "The Moon": {
-        "number": 18,
-        "keywords": ["illusion", "intuition", "fear"],
-        "description": "The Moon signifies intuition, illusions, and uncertainty. It suggests trusting your instincts amidst confusion."
-    },
-    "The Sun": {
-        "number": 19,
-        "keywords": ["positivity", "success", "joy"],
-        "description": "The Sun symbolizes positivity, success, and vitality. It encourages optimism and embracing the warmth of life."
-    },
-    "Judgement": {
-        "number": 20,
-        "keywords": ["rebirth", "reflection", "awakening"],
-        "description": "Judgement represents self-evaluation, awakening, and transformation. It calls for reflection and embracing a new phase."
-    },
-    "The World": {
-        "number": 21,
-        "keywords": ["completion", "fulfillment", "achievement"],
-        "description": "The World signifies the completion of a journey, fulfillment, and ultimate success. It encourages celebration and gratitude."
-    }
-}
