+++ app/static/js/websocket_chat.js
@@ -5,68 +5,67 @@ document.addEventListener("DOMContentLoaded", () => {
     const sidebar = document.getElementById("sidebar");
     const menuIcon = document.getElementById("menu-icon");
 
-    // WebSocket 연결 설정
-    const socket = io.connect('http://localhost:5000');  
+    // WebSocket 자동 연결 방지
+    const socket = io("http://localhost:5000", { autoConnect: false });
 
-
-    // 사이드바 상태를 저장
     let isSidebarOpen = false;
-    // 초기 토픽 값
     let selectedTopic = null;
-    
-    // 사이드바 열고 닫기 함수
+    let isSocketConnected = false;  // WebSocket이 연결되었는지 확인하는 변수
+
+    // 사이드바 열고 닫기
     function toggleSidebar() {
         sidebar.classList.toggle("active");
         isSidebarOpen = !isSidebarOpen;
     }
-    // 메뉴 아이콘 클릭 -> 사이드바 열고 닫기
     menuIcon.addEventListener("click", toggleSidebar);
 
-
-    // 토픽 선택 함수
-    function selectTopic(event) {
-        // 클릭된 버튼이 토픽 버튼인 경우에만 처리
-        if (event.target.classList.contains("topic-btn")) {
-            selectedTopic = event.target.textContent;  // 선택된 토픽 설정
-            console.log("선택된 토픽:", selectedTopic);
-
-            // 선택된 토픽 버튼에 스타일 추가 (활성화된 토픽 버튼 스타일링)
-            const topicButtons = document.querySelectorAll('.topic-btn');
-            topicButtons.forEach(btn => btn.classList.remove('active')); // 기존 버튼 스타일 제거
-            event.target.classList.add('active'); // 클릭된 버튼에 active 클래스 추가
+    // ✅ 특정 이벤트에서 WebSocket 연결 (예: 토픽 선택 시)
+    function connectSocketIfNeeded() {
+        if (!isSocketConnected) {
+            socket.connect();  // WebSocket 연결
+            console.log("WebSocket 연결됨!");
+            isSocketConnected = true;  // 연결 상태 업데이트
         }
     }
 
-    // 토픽 버튼 클릭 시 토픽 선택
-    const topicButtons = document.querySelectorAll('.topic-btn');
-    topicButtons.forEach(btn => {
-        btn.addEventListener('click', selectTopic);
-    });
+    // 토픽 선택 함수
+    function selectTopic(event) {
+        if (event.target.classList.contains("topic-btn")) {
+            selectedTopic = event.target.textContent;
+            console.log("선택된 토픽:", selectedTopic);
 
-    
-    // 서버에서 WebSocket을 통해 메시지를 받으면 화면에 표시
+            // 기존 버튼 스타일 제거 후, 선택한 버튼 활성화
+            document.querySelectorAll('.topic-btn').forEach(btn => btn.classList.remove('active'));
+            event.target.classList.add('active');
+
+            // ✅ 토픽 선택 시 WebSocket 연결 실행
+            connectSocketIfNeeded();
+        }
+    }
+
+    // 토픽 버튼 클릭 시 이벤트 리스너 추가
+    document.querySelectorAll('.topic-btn').forEach(btn => btn.addEventListener('click', selectTopic));
+
+    // 서버에서 WebSocket 메시지를 받으면 화면에 추가
     socket.on('new_message', function(data) {
         const messageDiv = document.createElement("div");
         messageDiv.className = `message ${data.sender}`;
         messageDiv.textContent = `${data.message}`;
         chatBox.appendChild(messageDiv);
-
-        // 스크롤을 아래로 이동 (최신 메시지 보이도록)
         chatBox.scrollTop = chatBox.scrollHeight;
     });
 
     // 메시지 전송 함수 (WebSocket 사용)
     function sendMessage() {
-        const userMessage = messageInput.value.trim();
-
-        if (userMessage && selectedTopic) {
-            // WebSocket을 통해 서버로 메시지 전송
-            socket.emit('send_message', { topic: selectedTopic, text: userMessage });
-
-            // 메시지 입력 필드 초기화
-            messageInput.value = '';
-        } else {
+        if (!selectedTopic) {
             alert("토픽을 먼저 선택해주세요!");
+            return;
+        }
+
+        const userMessage = messageInput.value.trim();
+        if (userMessage) {
+            socket.emit('send_message', { topic: selectedTopic, text: userMessage });
+            messageInput.value = '';
         }
     }
 
@@ -76,9 +75,8 @@ document.addEventListener("DOMContentLoaded", () => {
     // 'Enter' 키를 눌렀을 때 메시지 전송
     messageInput.addEventListener("keydown", (event) => {
         if (event.key === "Enter") {
-            event.preventDefault();  // 기본 Enter 동작(줄 바꿈) 방지
+            event.preventDefault();
             sendMessage();
         }
     });
 });
-
